requirements.txt
astroid==3.3.8
audioread==3.0.1
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
fonttools==4.56.0
isort==6.0.1
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pygame==2.6.1
pylint==3.3.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
SimpleGUICS2Pygame==2.1.0
six==1.17.0
tomlkit==0.13.2
